% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureSimulation.R
\name{betaSim}
\alias{betaSim}
\title{Beta Mixture Sampling}
\usage{
betaSim(ref_betamatrix, ref_phenotype)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}
}
\value{
The simulated purified profiles.
}
\description{
Beta Mixture Sampling simulation
}
