% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureSimulation.R
\name{betaSim_purifiedProfiles}
\alias{betaSim_purifiedProfiles}
\title{Purified profiles sampling in Beta Mixture Simulation.}
\usage{
betaSim_purifiedProfiles(ref_betamatrix, ref_phenotype, n = 20)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}

\item{n}{The number of simulated purfied cell profiles. Default value: 20.}
}
\value{
The simulated purified profiles.
}
\description{
Purified profiles sampling in Beta Mixture Simulation.
}
