% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReference.R
\name{build_reference_450k_neuron}
\alias{build_reference_450k_neuron}
\title{Build the reference library for 450k arrays of brain tissue}
\usage{
build_reference_450k_neuron()
}
\value{
A list of beta value reference matrix (ref_betamatrix) and reference cell types (ref_phenotype).
}
\description{
Build the reference library for 450k arrays of brain tissues (DLPFC) to include "NeuN_neg", "NeuN_pos".
}
