% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReference.R
\name{build_reference_EPIC}
\alias{build_reference_EPIC}
\title{Build the reference library for EPIC arrays}
\usage{
build_reference_EPIC(extend = TRUE)
}
\arguments{
\item{extend}{If TRUE, generate the extended reference library; otherwise, generate the reference library of
only 6 immune cell types. Default value is TRUE.}
}
\value{
A list of beta value reference matrix (ref_betamatrix) and reference cell types (ref_phenotype).
}
\description{
Build the reference library for EPIC arrays to include "CD8T", "CD4T", "NK", "Bcell", "Mono", "Neu",
And extend the reference library for 450k to include "Epithelial".
}
