% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_input}
\alias{check_input}
\title{Check input matrices}
\usage{
check_input(meth, unmeth)
}
\arguments{
\item{meth}{matrix with quantities of methylated sites}

\item{unmeth}{matrix with quantities of methylated sites}
}
\description{
Check input matrices
}
