% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_input_beta}
\alias{check_input_beta}
\title{Check input beta matrix}
\usage{
check_input_beta(beta_matrix)
}
\arguments{
\item{beta_matrix}{beta matrix}
}
\value{
same matrix, but fixed
}
\description{
Check input beta matrix
}
