% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compare_results}
\alias{compare_results}
\title{Show scatterplot that compares two results}
\usage{
compare_results(res1, res2)
}
\arguments{
\item{res1}{result1}

\item{res2}{result2}
}
\description{
Show scatterplot that compares two results
}
