% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create_beta}
\alias{create_beta}
\title{Create beta matrix from meth and unmeth matrices}
\usage{
create_beta(meth, unmeth)
}
\arguments{
\item{meth}{}

\item{unmeth}{}
}
\value{
beta matrix
}
\description{
Create beta matrix from meth and unmeth matrices
}
