% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{deconvolute_combined}
\alias{deconvolute_combined}
\title{Run selected set of methods and average results}
\usage{
deconvolute_combined(
  methyl_set,
  array = c("450k", "EPIC"),
  methods,
  scale_results = FALSE,
  ...
)
}
\arguments{
\item{methyl_set}{A minfi MethylSet}

\item{array}{type of methylation array that was used. possible options are '450k' and 'EPIC'}

\item{methods}{list of methods (>1) that will be applied to the methyl set}

\item{scale_results}{Whether the deconvolution results should be rescaled.
Negative values will be set to 0, and the estimates will be normalized to sum to 1 per sample.
Defaults to FALSE.}

\item{...}{Additional parameters, passed to the algorithm used. See individual method documentations for details.}
}
\value{
dataframe with results of all selected methods as well as the combined estimates
}
\description{
Run selected set of methods and average results
}
\examples{

ex_data <- minfiData::MsetEx

result <- deconvolute_combined(ex_data, methods=c('epidish','houseman'))
}
