% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolution}
\usage{
deconvolve(
  methyl_set,
  method = deconvolution_methods,
  scale_results = FALSE,
  ...
)
}
\arguments{
\item{methyl_set}{A minfi MethylSet}

\item{method}{A string specifying the method. Supported methods are 'epidish', 'houseman', 'methylcc', 'methylresolver', 'methatlas'}

\item{scale_results}{Whether the deconvolution results should be rescaled.
Negative values will be set to 0, and the estimates will be normalized to sum to 1 per sample.
Defaults to FALSE.}

\item{...}{Additional parameters, passed to the algorithm used. See individual method documentations for details.}
}
\value{
A matrix with the probabilities of each cell-type for each individual. Rows are
individuals, columns are cell types.
}
\description{
Deconvolution
}
\examples{

ex_data <- minfiData::MsetEx

result <- deconvolve(ex_data, method='epidish')
}
