% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{deconvolve_combined}
\alias{deconvolve_combined}
\title{Run selected set of methods and aggregate results}
\usage{
deconvolve_combined(
  methyl_set,
  array = c("450k", "EPIC"),
  methods,
  aggregation_method = "mean",
  scale_results = FALSE
)
}
\arguments{
\item{methyl_set}{A minfi MethylSet}

\item{array}{type of methylation array that was used. possible options are '450k' and 'EPIC'}

\item{methods}{list of methods (>1) that will be applied to the methyl set}

\item{aggregation_method}{how should deconvolution estimates be combined (default is mean)}

\item{scale_results}{Whether the deconvolution results should be rescaled.
Negative values will be set to 0, and the estimates will be normalized to sum to 1 per sample.
Defaults to FALSE.}
}
\value{
dataframe with results of all selected methods as well as the combined estimates
}
\description{
Run selected set of methods and aggregate results
}
