% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvAlgorithms.R
\name{enrichment_score}
\alias{enrichment_score}
\title{Enrichment score-based algorithm}
\usage{
enrichment_score(
  betamatrix,
  ref_betamatrix,
  ref_phenotype,
  method = "MCP-counter",
  pv = 1e-08,
  MaxDMRs = 100
)
}
\arguments{
\item{betamatrix}{Beta value matrix of methylation array for mixture samples.}

\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}

\item{method}{The enrichment score-based deconvolution algorithm such as "MCP-counter", "ssGSEA" and "ESTIMATE",
which are adapted from gene-expression based algorithm. Default value: "MCP-counter".}

\item{pv}{The p-value threshold with default value as 1e-8}

\item{MaxDMRs}{The number of probes selected with default value as 100.}
}
\value{
The matrix of deconvolution results with rows as mixture samples and columns as cell types.
}
\description{
CBS deconvolution algorithm
}
