% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixtureSimulation.R
\name{GaussianSim_mixtureProfiles}
\alias{GaussianSim_mixtureProfiles}
\title{Gaussian Mixture Simulation}
\usage{
GaussianSim_mixtureProfiles(nonimmune_level = 1, purified_datasets_sim, n = 30)
}
\arguments{
\item{nonimmune_level}{The levels are 1,2,3,4,5. Default value is 1. 1: No non-immune component; level 2: the non-immune proportion is 0.1-0.2;
level 3: the non-immune proportion is 0.2-0.5; level 4: the non-immune proportion is 0.5-0.8; level 5: the non-immune proportion is 0.8-0.9.}

\item{purified_datasets_sim}{The reference matrix ref_betamatrix.}

\item{n}{The number of simulated proportions. Default value: 30.}
}
\value{
The simulated mixture profiles.
}
\description{
Gaussian Mixture Simulation to generate the mixture profiles with true/known proportions.
}
