% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{MethylDeconv_pipeline}
\alias{MethylDeconv_pipeline}
\title{Deconvolution Pipeline}
\usage{
MethylDeconv_pipeline(
  input_methyl,
  normalized = FALSE,
  array_type = "450k",
  feature_selection = "oneVsAllttest",
  deconv_algorithm = "Houseman",
  tissue = "general",
  extend_reference = TRUE,
  custom_probes = NULL
)
}
\arguments{
\item{input_methyl}{Methylation array for mixture samples. The rows correpsond to the probes/CpGs;
the columns correspond to the samples.}

\item{normalized}{If TRUE, the methylation array is the beta value matrix. Default value: FALSE.}

\item{array_type}{The array type can be 450k or EPIC. Default value: "450k".}

\item{feature_selection}{Different feature selection methods such as "oneVsAllttest" (Default value), "oneVsAllLimma",
"pairwiseLimma", "pairwiseGlmnet", "multiGlmnet", "glmnetpreselect", "RFpreselect", "OptVariables".}

\item{deconv_algorithm}{Different deconvolution algorithms such as "Houseman" (default value), "RPC", "CBS", "MethylResolver",
"MCP-counter", "ssGSEA", "ESTIMATE".}

\item{tissue}{Different tissue types such as "general" (default value) which correspond to blood tissue and general epithelial
tissue; "brain" correpsonds to brain tissues.}

\item{extend_reference}{Default: TRUE. If TRUE, extend the reference library.}

\item{custom_probes}{Default: NULL. The user-defined probe lists.}
}
\value{
The deconvolution results (cell type proportions) with rows as the samples and the columns as cell types.
}
\description{
Deconvolution Pipeline
}
