% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refCompTableProbeSelection.R
\name{ref_compTable}
\alias{ref_compTable}
\title{Average reference profiles over each cell type}
\usage{
ref_compTable(ref_betamatrix, ref_phenotype)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}
}
\value{
The data maxtirx of average reference profiles (each column corresponds to each cell type),
except for the first column (Fstat) and the last three columns (low, high, range) which are the summary statistics.
}
\description{
Compute the average reference profiles for each cell type, using the reference matrix of purified samples
from different cell types (phenotype).
}
