% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refCompTableProbeSelection.R
\name{ref_probe_selection_HighVar}
\alias{ref_probe_selection_HighVar}
\title{Feature selection based on variability}
\usage{
ref_probe_selection_HighVar(ref_betamatrix, ranks = 601:1200)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ranks}{The top number of most variable probes. Default value is 601:1200,
which corresponds the second 600 highest variable probes.}
}
\value{
A vector of the selected probes.
}
\description{
Select the most variable probes across all samples in the reference matrix
}
