% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refCompTableProbeSelection.R
\name{ref_probe_selection_multiclassRF_cv}
\alias{ref_probe_selection_multiclassRF_cv}
\title{Multi-class RF modeling feature selection}
\usage{
ref_probe_selection_multiclassRF_cv(
  ref_betamatrix,
  ref_phenotype,
  nCores = 4,
  reps.resamp = 5,
  reps.repeats = 3,
  tune_grid = 1:30
)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}
}
\value{
A vector of the selected probes.
}
\description{
Select features based on multi-class Random Forest modeling on the reference matrix of all probes.
}
