% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refCompTableProbeSelection.R
\name{ref_probe_selection_oneVsAllLimma}
\alias{ref_probe_selection_oneVsAllLimma}
\title{One-vs-All moderated t test feature selection}
\usage{
ref_probe_selection_oneVsAllLimma(
  ref_betamatrix,
  ref_phenotype,
  probeSelect = "both",
  FDR = 1e-06,
  MaxDMRs = 100
)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}

\item{probeSelect}{The selection can be "any" or "both". If "any", the function selects top probes regardless of
up-regulation or down-regulation. If "both", half of top probes are picked up from the up-regulated probes while
the other half of the top probes are picked up from the down-regulated probes. Default value is "both".}

\item{FDR}{The fdr threshold with default value as 1e-6}

\item{MaxDMRs}{The number of probes selected with default value as 100.}
}
\value{
A vector of the selected probes.
}
\description{
The moderated One-vs-All t test feature selection based on the reference matrix.
}
