% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refCompTableProbeSelection.R
\name{ref_probe_selection_oneVsAllttest}
\alias{ref_probe_selection_oneVsAllttest}
\title{One-vs-All t test feature selection}
\usage{
ref_probe_selection_oneVsAllttest(
  ref_betamatrix,
  ref_phenotype,
  probeSelect = "both",
  pv = 1e-08,
  MaxDMRs = 100
)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}

\item{probeSelect}{The selection can be "any" or "both". If "any", the function selects top probes regardless of
up-regulation or down-regulation. If "both", half of top probes are picked up from the up-regulated probes while
the other half of the top probes are picked up from the down-regulated probes. Default value is "both".}

\item{pv}{The p-value threshold with default value as 1e-8}

\item{MaxDMRs}{The number of probes selected with default value as 100.}
}
\value{
A vector of the selected probes.
}
\description{
The One-vs-All t test feature selection based on the reference matrix.
}
