% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refCompTableProbeSelection.R
\name{ref_probe_selection_pairwiseGlmnet_cv}
\alias{ref_probe_selection_pairwiseGlmnet_cv}
\title{Pairwise elastic net modeling feature selection}
\usage{
ref_probe_selection_pairwiseGlmnet_cv(
  ref_betamatrix,
  ref_phenotype,
  nCores = 4,
  reps.resamp = 5,
  reps.repeats = 3
)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}

\item{nCores}{Number of parallel cores. Default value is 4.}

\item{reps.resamp}{The number of folds in k-fold cross-validation. Default value is 5.}

\item{reps.repeats}{The number of complete sets of folds to compute. Default value is 3.}
}
\value{
A vector of the selected probes.
}
\description{
Select the non-zero features from the pairwise elastic net modeling on the reference matrix.
}
