% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refCompTableProbeSelection.R
\name{ref_probe_selection_pairwiseLimma}
\alias{ref_probe_selection_pairwiseLimma}
\title{Pairwise moderated t test feature selection}
\usage{
ref_probe_selection_pairwiseLimma(
  ref_betamatrix,
  ref_phenotype,
  FDR = 0.01,
  deltaBeta = 0.2,
  MaxDMRs = 100
)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}

\item{FDR}{The fdr threshold with default value as 0.01}

\item{deltaBeta}{The threshold for the absolute mean difference. Default value is 0.2.}

\item{MaxDMRs}{The number of probes selected with default value as 100.}
}
\value{
A vector of the selected probes.
}
\description{
The pairwise moderated t test feature selection based on the reference matrix.
}
