% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refCompTableProbeSelection.R
\name{ref_probe_selection_twoStage}
\alias{ref_probe_selection_twoStage}
\title{Two-stage feature selection}
\usage{
ref_probe_selection_twoStage(
  ref_betamatrix,
  ref_phenotype,
  preselect = 300,
  ml_model = "elastic net"
)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}

\item{preselect}{The number of top features per cell type selected from one-vs-all t tests. The default value is 300.}

\item{ml_model}{The machine learning model for feature selection in the second stage. The default value is "elastic net",
which correpsonds to selecting the non-zero features from multi-class elastic net modeling on the reference matrix. Otherwise,
if the parameter value is "RF", the model selection is based on the important variables learnt from multi-class Random forest modeling;
if the parameter value is "rfe", it selects features based on recursive feature elimination algorithm and
a Random Forest algorithm is used on each iteration to evaluate the model.}
}
\value{
Model class.
}
\description{
Select features in two stages: firstly, select top features from one-vs-all t test;
secondly, select the features with machine learning modeling.
}
