% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{results_aggregated_boxplot}
\alias{results_aggregated_boxplot}
\title{Function to plot the results from \code{deconvolute_combined()} as boxplots for each celltype and celltype}
\usage{
results_aggregated_boxplot(result)
}
\arguments{
\item{result}{result from \code{deconvolute_combined()}}
}
\description{
Function to plot the results from \code{deconvolute_combined()} as boxplots for each celltype and celltype
}
