% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{results_barplot}
\alias{results_barplot}
\title{Function to plot the results from \code{deconvolute()}  as barplots for each sample}
\usage{
results_barplot(result)
}
\arguments{
\item{result}{result from \code{deconvolute()}}
}
\description{
Function to plot the results from \code{deconvolute()}  as barplots for each sample
}
