% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{results_boxplot}
\alias{results_boxplot}
\title{Function to plot the results from \code{deconvolute()}  as boxplots for each celltype}
\usage{
results_boxplot(result)
}
\arguments{
\item{result}{result from \code{deconvolute()}}
}
\description{
Function to plot the results from \code{deconvolute()}  as boxplots for each celltype
}
