% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvAlgorithms.R
\name{RPC}
\alias{RPC}
\title{RPC algorithm}
\usage{
RPC(betamatrix, compTable, probes_select)
}
\arguments{
\item{betamatrix}{Beta value matrix of methylation array for mixture samples.}

\item{compTable}{Average reference profiles over each cell type.}

\item{probes_select}{Probes selected.}
}
\value{
The matrix of deconvolution results with rows as mixture samples and columns as cell types.
}
\description{
RPC deconvolution algorithm
}
