% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_deconvolutions}
\alias{run_deconvolutions}
\title{Main function}
\usage{
run_deconvolutions(
  obj = NA,
  meth = NA,
  unmeth = NA,
  meta = NA,
  condition = NA,
  seed = 1
)
}
\value{
A String
}
\description{
Main function
}
