% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidish.R
\name{run_epidish_cp}
\alias{run_epidish_cp}
\title{Title}
\usage{
run_epidish_cp(value_matrix)
}
\arguments{
\item{value_matrix}{matrix containing methylation profiles}
}
\value{
Cell Type proportion matrix
}
\description{
Title
}
