% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowsortedbloodepic.R
\name{run_flowsortedbloodepic}
\alias{run_flowsortedbloodepic}
\title{run FlowSortedBloodEPIC deconvolution}
\usage{
run_flowsortedbloodepic(obj, method = "preprocessQuantile")
}
\arguments{
\item{obj}{minfi::MethylSet, minfi::GenomicMethylSet or minfi::RGChannelSet}

\item{method}{"preprocessQuantile" or "preprocessNoob"}
}
\value{
Cell Type proportion matrix
}
\description{
run FlowSortedBloodEPIC deconvolution
}
