% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowsortedbloodepic.R
\name{run_flowsortedbloodepic_raw}
\alias{run_flowsortedbloodepic_raw}
\title{run FlowSortedBloodEPIC deconvolution using raw files}
\usage{
run_flowsortedbloodepic_raw(meth, unmeth)
}
\arguments{
\item{meth}{raw methylation data (meth)}

\item{unmeth}{raw methylation data (unmeth)}
}
\value{
Cell Type proportion matrix
}
\description{
run FlowSortedBloodEPIC deconvolution using raw files
}
