% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseman.R
\name{run_houseman}
\alias{run_houseman}
\title{Run the improved Houseman method}
\usage{
run_houseman(
  methyl_set,
  array = c("450k", "EPIC"),
  compositeCellType = c("Blood", "CordBloodCombined", "CordBlood", "CordBloodNorway",
    "CordTissueAndBlood", "DLPFC"),
  processMethod = "preprocessQuantile",
  probeSelect = c("IDOL", "both", "any"),
  cellTypes = c("CD8T", "CD4T", "NK", "Bcell", "Mono", "Neu"),
  referencePlatform = c("IlluminaHumanMethylationEPIC", "IlluminaHumanMethylation450k",
    "IlluminaHumanMethylation27k"),
  referenceset = NULL,
  CustomCpGs = NULL,
  meanPlot = FALSE,
  verbose = TRUE,
  lessThanOne = FALSE,
  cellCounts = NULL,
  ...
)
}
\arguments{
\item{methyl_set}{A minfi MethylSet}

\item{array}{type of methylation array that was used. possible options are '450k' and 'EPIC'}

\item{compositeCellType}{Which composite cell type is being deconvoluted. Should be one of "Blood", "CordBloodCombined", "CordBlood", "CordBloodNorway", "CordTissueAndBlood", or "DLPFC". See details for preferred approaches.}

\item{processMethod}{Joint normalization/background correction for user and reference data. For MethylSet objects only "preprocessQuantile" is available. Set it to any minfi preprocessing function as a character if you want to override it, like "preprocessFunnorm"}

\item{probeSelect}{How should probes be selected to distinguish cell types? Options include: 1) "IDOL", (default) for using a customized set of probes obtained from IDOL optimization, available for Blood and Umbilical Cord Blood 2) "both", which selects an equal number (50) of probes (with F-stat p-value < 1E-8) with the greatest magnitude of effect from the hyper- and hypo-methylated sides, and 3) "any", which selects the 100 probes (with F-stat p-value < 1E-8) with the greatest magnitude of difference regardless of direction of effect. This according to minfi algorithm. Default input "auto" in minfi will use "any" for cord blood and "both" otherwise. Please see references for more details.}

\item{cellTypes}{A vector of length K that contains the cell type names. Default: c("CD8T", "CD4T", "NK", "Bcell", "Mono", "Neu"). Please notice that this library use Neutrophils instead of Granulocytes. See details for your library.}

\item{referencePlatform}{The platform for the reference dataset; options include c("IlluminaHumanMethylation450k", "IlluminaHumanMethylationEPIC" (default), "IlluminaHumanMethylation27k"). If the input rgSet belongs to another platform, it will be converted using minfi function convertArray. Should not be changed by the user.}

\item{referenceset}{It is NULL by default. A custom reference RGChannelSet object (in quotes) if it is not a package installed. This option also allows the user to perform the deconvolution in closed computing clusters without internet access to ExperimentHub. For that download and save the reference and input the resulting object here. If using an installed reference package set to NULL.}

\item{CustomCpGs}{a custom vector of probe names for cell deconvolution. For custom lists it should be a vector object (no quotes).}

\item{meanPlot}{Whether to plots the average DNA methylation across the cell-type discriminating probes within the mixed and sorted samples.}

\item{verbose}{Should the function be verbose?}

\item{lessThanOne}{Should the predictions be constrained to exactly one, in minfi default is FALSE, now you can select the option}

\item{cellCounts}{If cell counts are available (CBC, of flow sort) add a vector of lenght equal to the samples being deconvolved}

\item{...}{Other arguments for preprocessQuantile or other normalizations}

\item{returnAll}{Should the composition table and the normalized user supplied data be return? Default is False.}
}
\description{
Run the improved Houseman method
}
