% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylcc.R
\name{run_methylcc_raw}
\alias{run_methylcc_raw}
\title{run methylCC deconvolution using raw files}
\usage{
run_methylcc_raw(meth, unmeth, seed = 1)
}
\arguments{
\item{meth}{raw methylation data (meth)}

\item{unmeth}{raw methylation data (unmeth)}

\item{seed}{numeric}
}
\value{
Cell Type proportion matrix
}
\description{
run methylCC deconvolution using raw files
}
