% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylResolver.R
\name{run_methylresolver}
\alias{run_methylresolver}
\title{Run MethylResolver}
\usage{
run_methylresolver(
  beta_matrix,
  doPar = F,
  numCores = 1,
  alpha = seq(0.5, 0.9, by = 0.05),
  absolute = TRUE,
  purityModel = MethylResolver::RFmodel,
  seed = 1
)
}
\arguments{
\item{beta_matrix}{a beta matrix with CpGs in rows and samples in columns}

\item{doPar}{Whether to use parallel processing to speed up the deconvolution computation if many samples are present. Default is FALSE.}

\item{numCores}{Number of cores used for parallel processing to speed up the deconvolution of many samples. Requires doPar = TRUE. Default is 1. numCores = "auto" is max number of cores available minus one.}

\item{alpha}{Set the alpha parameter for LTS deconvolution. This is the fraction of optimal CpGs from the signature matrix which are used for deconvolution. Must be between 0 and 1. Users can specify a vector or a single number. If a vector is specified, a grid search of the values is conducted and the alpha value that results in the lowest RMSE between the original and reconstructed mixture is selected. Default is seq(0.5,0.9,by = 0.05).}

\item{absolute}{Whether to compute tumor purity and absolute cell type fractions. Default is TRUE.}

\item{purityModel}{Random Forest model to predict mixture purity (unknown content) which allows the calculation of absolute cell type fractions. Required if absolute is TRUE. Default is our RF model trained on the consensus purity estimate (CPE) using TCGA data.}

\item{seed}{fixed seed to account for RNG influences}
}
\description{
Run MethylResolver
}
