% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tca.R
\name{run_tca}
\alias{run_tca}
\title{run TCA}
\usage{
run_tca(value_matrix, meta_data, condition, epidish_output)
}
\arguments{
\item{value_matrix}{matrix containing methylation profiles}

\item{meta_data}{matrix containing information about samples in 'value_matrix'}

\item{condition}{collumn name of 'meta_data' either complex or numeric}
}
\value{
Cell Type proportion matrix
}
\description{
run TCA
}
