% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refCompTableProbeSelection.R
\name{up_probes_oneVsAllttest_celltype}
\alias{up_probes_oneVsAllttest_celltype}
\title{One-vs-All t test feature selection}
\usage{
up_probes_oneVsAllttest_celltype(
  ref_betamatrix,
  ref_phenotype,
  pv = 1e-08,
  MaxDMRs = 100
)
}
\arguments{
\item{ref_betamatrix}{The reference matrix ref_betamatrix.}

\item{ref_phenotype}{The cell type information for the reference matrix.}

\item{pv}{The p-value threshold with default value as 1e-8}

\item{MaxDMRs}{The number of probes selected with default value as 100.}
}
\value{
A vector of the selected probes.
}
\description{
The One-vs-All t test feature selection based on the reference matrix.
}
