% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visualize_result_bar}
\alias{visualize_result_bar}
\title{Function to visualize the results from \code{deconvolute()} as a stacked bar plot.}
\usage{
visualize_result_bar(result, rename = F)
}
\arguments{
\item{result}{result object from \code{deconvolute()}}

\item{rename}{if the cell types should be filtered and renamed for better comparison (default: FALSE).
Might be most useful for MethAtlas.}
}
\description{
Function to visualize the results from \code{deconvolute()} as a stacked bar plot.
}
