% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visualize_results}
\alias{visualize_results}
\title{Function to plot the results from \code{deconvolute()} or \code{run_all_methods()} as boxplots.}
\usage{
visualize_results(result)
}
\arguments{
\item{result}{result from \code{deconvolute()} or \code{run_all_methods()}}
}
\description{
Function to plot the results from \code{deconvolute()} or \code{run_all_methods()} as boxplots.
}
