% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create_beta}
\alias{create_beta}
\title{Create beta matrix from a methylset matrices}
\usage{
create_beta(methyl_set)
}
\arguments{
\item{methyl_set}{}
}
\value{
beta matrix
}
\description{
Create beta matrix from a methylset matrices
}
