% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rename_cell_types}
\alias{rename_cell_types}
\title{Rename Cell Types in Deconvolution Results}
\usage{
rename_cell_types(input_celltypes)
}
\arguments{
\item{input_celltypes}{list of cell types}
}
\value{
a new list of cell types with controlled vocabulary
}
\description{
Renames a set of cell type names to standardized names.
It uses predefined mappings to replace specific cell type names with more
consistent labels (e.g., "CD8T" to "T cell CD8+"). Unrecognized names are labeled as "other".
}
