% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidish.R
\name{run_epidish}
\alias{run_epidish}
\title{run EpiDISH}
\usage{
run_epidish(value_matrix)
}
\arguments{
\item{value_matrix}{matrix containing methylation profiles}
}
\value{
3x Cell Type proportion matrices
}
\description{
run EpiDISH
}
