% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowsortedbloodepic.R
\name{run_flowsortedbloodepic}
\alias{run_flowsortedbloodepic}
\title{run FlowSortedBloodEPIC deconvolution}
\usage{
run_flowsortedbloodepic(methyl_set)
}
\arguments{
\item{methyl_set}{minfi::MethylSet}
}
\value{
Cell Type proportion matrix
}
\description{
run FlowSortedBloodEPIC deconvolution
}
