% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylcc.R
\name{run_methylcc}
\alias{run_methylcc}
\title{run methylCC deconvolution}
\usage{
run_methylcc(genomic_methyl_set, seed = 1)
}
\arguments{
\item{genomic_methyl_set}{minfi::MethylSet or minfi::GenomicMethylSet}

\item{seed}{numeric}
}
\value{
Cell Type proportion matrix
}
\description{
run methylCC deconvolution
}
